+++ build_config.rb
@@ -15,6 +15,7 @@ def rpg_maker_gems(conf)
   conf.gem "#{MRUBY_ROOT}/../../mruby-rgss"
   conf.gem "#{MRUBY_ROOT}/../../mruby-rpg2k"
   conf.gem "#{MRUBY_ROOT}/../../mruby-rpgxp"
+  conf.gem "#{MRUBY_ROOT}/../../mruby-mvjs"
 end
 
 # When targeting Emscripten the host build only exists to produce the `mrbc`